+++ build.gradle.kts
@@ -26,38 +26,12 @@
 
 @file:Suppress("RemoveRedundantQualifierName") // To prevent IDEA replacing FQN imports.
 
-import Build_gradle.Subproject
-import io.spine.internal.dependency.CheckerFramework
-import io.spine.internal.dependency.Coroutines
-import io.spine.internal.dependency.ErrorProne
-import io.spine.internal.dependency.FindBugs
-import io.spine.internal.dependency.Grpc
-import io.spine.internal.dependency.Guava
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Kotlin
-import io.spine.internal.dependency.Protobuf
-import io.spine.internal.dependency.Spine
-import io.spine.internal.dependency.Truth
-import io.spine.internal.gradle.VersionWriter
-import io.spine.internal.gradle.checkstyle.CheckStyleConfig
-import io.spine.internal.gradle.excludeProtobufLite
-import io.spine.internal.gradle.forceVersions
-import io.spine.internal.gradle.github.pages.updateGitHubPages
-import io.spine.internal.gradle.javac.configureErrorProne
-import io.spine.internal.gradle.javac.configureJavac
-import io.spine.internal.gradle.javadoc.JavadocConfig
-import io.spine.internal.gradle.kotlin.applyJvmToolchain
-import io.spine.internal.gradle.kotlin.setFreeCompilerArgs
-import io.spine.internal.gradle.publish.IncrementGuard
 import io.spine.internal.gradle.publish.PublishingRepos
 import io.spine.internal.gradle.publish.spinePublishing
 import io.spine.internal.gradle.report.coverage.JacocoConfig
 import io.spine.internal.gradle.report.license.LicenseReporter
 import io.spine.internal.gradle.report.pom.PomGenerator
 import io.spine.internal.gradle.standardToSpineSdk
-import io.spine.internal.gradle.testing.configureLogging
-import io.spine.internal.gradle.testing.registerTestTasks
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 buildscript {
     standardSpineSdkRepositories()
@@ -79,14 +53,14 @@ buildscript {
 }
 
 plugins {
-    `java-library`
-    kotlin("jvm")
     idea
-    errorprone
     jacoco
     `project-report`
     `gradle-doctor`
 }
+JacocoConfig.applyTo(project)
+PomGenerator.applyTo(project)
+LicenseReporter.mergeAllReports(project)
 
 spinePublishing {
     modules = setOf(
@@ -103,13 +77,7 @@ spinePublishing {
 }
 
 allprojects {
-    apply {
-        plugin("jacoco")
-        plugin("idea")
-        plugin("project-report")
-        from("$rootDir/version.gradle.kts")
-    }
-
+    apply(from = "$rootDir/version.gradle.kts")
     group = "io.spine.tools"
     version = extra["versionToPublish"]!!
 
@@ -117,130 +85,5 @@ allprojects {
 }
 
 subprojects {
-    applyPlugins()
-    addDependencies()
-    forceConfigurations()
-
-    val javaVersion = JavaLanguageVersion.of(11)
-    configureJava(javaVersion)
-    configureKotlin(javaVersion)
-
-    configureTests()
-
-    configureGitHubPages()
-    configureTaskDependencies()
+    apply(plugin = "module")
 }
-
-JacocoConfig.applyTo(project)
-PomGenerator.applyTo(project)
-LicenseReporter.mergeAllReports(project)
-
-typealias Subproject = Project
-
-fun Subproject.applyPlugins() {
-    apply {
-        plugin("java-library")
-        plugin("kotlin")
-        plugin("net.ltgt.errorprone")
-        plugin("pmd-settings")
-        plugin(Protobuf.GradlePlugin.id)
-        plugin("write-manifest")
-    }
-
-    apply<IncrementGuard>()
-    apply<VersionWriter>()
-
-    CheckStyleConfig.applyTo(project)
-    JavadocConfig.applyTo(project)
-    LicenseReporter.generateReportIn(project)
-}
-
-fun Subproject.addDependencies() {
-    dependencies {
-        errorprone(ErrorProne.core)
-
-        compileOnlyApi(FindBugs.annotations)
-        compileOnlyApi(CheckerFramework.annotations)
-        ErrorProne.annotations.forEach { compileOnlyApi(it) }
-
-        implementation(Guava.lib)
-
-        testImplementation(Guava.testLib)
-        JUnit.api.forEach { testImplementation(it) }
-        Truth.libs.forEach { testImplementation(it) }
-        testRuntimeOnly(JUnit.runner)
-    }
-}
-
-fun Subproject.forceConfigurations() {
-    with(configurations) {
-        forceVersions()
-        excludeProtobufLite()
-        all {
-            resolutionStrategy {
-                val spine = Spine(project)
-                @Suppress("DEPRECATION") /* Still need to force `Kotlin.stdLibJdk8` until full
-                  migration to Kotlin 1.8.0. */
-                force(
-                    JUnit.runner,
-                    spine.base,
-                    spine.validation.runtime,
-                    Grpc.stub,
-                    Coroutines.jdk8,
-                    Coroutines.core,
-                    Coroutines.bom,
-                    Coroutines.coreJvm,
-                    Kotlin.stdLibJdk8
-                )
-            }
-        }
-
-    }
-}
-
-fun Subproject.configureJava(javaVersion: JavaLanguageVersion) {
-    java {
-        toolchain.languageVersion.set(javaVersion)
-    }
-    tasks {
-        withType<JavaCompile>().configureEach {
-            configureJavac()
-            configureErrorProne()
-        }
-    }
-}
-
-fun Subproject.configureKotlin(javaVersion: JavaLanguageVersion) {
-    kotlin {
-        applyJvmToolchain(javaVersion.asInt())
-        explicitApi()
-    }
-
-    tasks {
-        withType<KotlinCompile>().configureEach {
-            kotlinOptions.jvmTarget = javaVersion.toString()
-            setFreeCompilerArgs()
-        }
-    }
-}
-
-fun Subproject.configureTests() {
-    tasks {
-        registerTestTasks()
-        test.configure {
-            useJUnitPlatform {
-                includeEngines("junit-jupiter")
-            }
-            configureLogging()
-        }
-    }
-}
-
-fun Subproject.configureGitHubPages() {
-    updateGitHubPages(Spine.DefaultVersion.javadocTools) {
-        allowInternalJavadoc.set(true)
-        rootFolder.set(rootDir)
-    }
-}
-
-
