+++ tool-base/build.gradle.kts
@@ -33,6 +33,7 @@ import io.spine.internal.dependency.JavaPoet
 import io.spine.internal.dependency.JavaX
 import io.spine.internal.dependency.Protobuf
 import io.spine.internal.dependency.Spine
+import io.spine.internal.gradle.protobuf.setup
 
 plugins {
     protobuf
@@ -95,6 +96,7 @@ protobuf {
             id(Grpc.ProtocPlugin.id)
             id(GrpcKotlin.ProtocPlugin.id)
         }
+        setup(generatedDir)
     }
 }
